+++ lib/CodeTransform/Adapter/WorseReflection/Transformer/CompleteConstructor.php
@@ -130,12 +130,40 @@ class CompleteConstructor implements Transformer
 
         foreach ($this->candidateClasses($source) as $class) {
             $constructMethod = $class->methods()->get('__construct');
+            $nonPromotedParameterNames = $this->getParentClassParamaterNames($class);
             foreach ($constructMethod->parameters()->notPromoted() as $parameter) {
+                if (in_array($parameter->name(), $nonPromotedParameterNames)) {
+                    continue;
+                }
                 $edits[] = TextEdit::create($parameter->position()->start()->toInt(), 0, sprintf('%s ', $this->visibility));
             }
         }
 
         return TextEdits::fromTextEdits($edits);
+    }
+
+    /**
+    * Get the names of the parent constructor to know which parameters should not be promoted.
+    * @return array<string>
+    */
+    private function getParentClassParamaterNames(ReflectionClass $class): array
+    {
+        $ancestor = $class->ancestors()->firstOrNull();
+        if ($ancestor === null) {
+            return [];
+        }
+
+        $constructMethod = $ancestor->methods()->get('__construct');
+        if ($constructMethod === null) {
+            return [];
+        }
+
+        $parameters = [];
+        foreach($constructMethod->parameters() as $parameter) {
+            $parameters[] = $parameter->name();
+        }
+
+        return $parameters;
     }
 
     /**